Add new features to the auction frontend and improve search params in backend
Creating new hooks (useParamsStore) and components (Filter, AppPagination, AuctionCard) that support pagination, filter and proper information display on each auction card. Added new dependencies to package.json for these functionalities. Also, made search params in the backend nullable for better flexibility in search queries. This is necessary due to the introduction of features in the frontend that require larger flexibility in the backend search parameters. Introduction of such features aims to enhance user experience and ease of use.