Refactor SearchParams class and cleanup codebase
Refactored the SearchParams class to not accept null values for better data consistency and less error-prone inputs. Removed unnecessary usages and references in the AuctionsController. The AuctionCreateFaultConsumer class is marked as abstract since it should not be instantiated. Explicitly ignored some compiler warnings where we are aware and accept the behavior. Renamed outbox to Outbox migration for naming consistency. Moved AuctionDeleted class to the correct namespace. Added qodana.yaml for upcoming Qodana code quality tool integration.