Add authentication and update search
Implemented user authentication using NextAuth.js. Added Login and UserActions components for sign-in functionality and user related options, respectively. Updated GetData in auctionActions to fetch data from API endpoint and added a method to update auction details. Made slight adjustments to how search results are sorted by adjusting the Search Controller parameters to also include model. The .dockerfile was created for building a Docker image for the web app. Additional setup files for MongoDB and navigation components were also added. Made changes in package.json to include "next-auth" for user authentication. This is to ensure users are able to log in and perform actions related to their profile, and to have a stable and secure means of authentication.