Add authentication test and token data retrieval functionality
A new file AuthTest.tsx was added to implement a test for authentication. A function doUpdate was added to the file to trigger the authentication process and show the result. The tokens.memory.access_token field in the route.ts file under the onToken callback and the token.access_token in next-auth.d.ts have been updated to help in capturing the token data. The page.tsx file was updated to display both the session and token data on the session page. The auctionAction.ts file was modified to use an authorization header when sending a PUT request with the token sent as a bearer token. A getTokenWorkaround function was added to authActions.ts to fetch the authorization token cookies.