Add AuctionSvcHttpClient and update app config and DB initializers
A new file for AuctionSvcHttpClient has been added to the SearchService, modifying the way data is brought from the AuctionService. GatewayService and SearchService app settings were also updated to handle the retrieval of the AuctionService base URL. Additionally, changes were made to Database initializers to enhance data seeding process in SearchService and AuctionService.